Problem: Agent verbosity is not properly set up.
Solution: Initialize verbose mode of the logger; use 'verbose' variable only when necessary.
(Since the agent does not have the config file, log config is hardwired for now. It will be made configurable after config file processing is available in fty-common, in order to avoid code duplication which would have to be resolved soon anyway.)

Signed-off-by: Jana Rapava <janarapava@eaton.com>